Fix Dropbox sync error
