+++ NAMESPACE
@@ -30,6 +30,7 @@ import(colorspace)
 import(data.table)
 import(dt.inflix)
 import(ggplot2)
+import(knitr)
 import(parallel)
 import(stringr)
 import(testthat)
