--- NAMESPACE (Andrew Rech's conflicted copy 2018-03-25)
@@ -1,57 +1,0 @@
-# Generated by roxygen2: do not edit by hand
-
-export(check_pred_tools)
-export(collate_netMHC)
-export(detect_mhc)
-export(extract_cDNA)
-export(garnish_fitness)
-export(garnish_jaffa)
-export(garnish_plot)
-export(garnish_predictions)
-export(garnish_summary)
-export(garnish_variants)
-export(get_metadata)
-export(get_pred_commands)
-export(get_snpeff)
-export(get_ss_str)
-export(list_mhc)
-export(make_BLAST_uuid)
-export(make_DAI_uuid)
-export(make_cDNA)
-export(make_nmers)
-export(merge_predictions)
-export(run_mhcflurry)
-export(run_mhcnuggets)
-export(run_netMHC)
-export(translate_cDNA)
-export(write_mhcnuggets_nmers)
-export(write_netmhc_nmers)
-import(colorspace)
-import(data.table)
-import(dt.inflix)
-import(ggplot2)
-import(knitr)
-import(parallel)
-import(stringr)
-import(testthat)
-importFrom(Biostrings,DNAString)
-importFrom(Biostrings,translate)
-importFrom(Rdpack,reprompt)
-importFrom(ShortRead,readFasta)
-importFrom(ShortRead,sread)
-importFrom(biomaRt,getBM)
-importFrom(biomaRt,getSequence)
-importFrom(biomaRt,useMart)
-importFrom(magrittr,"%$%")
-importFrom(magrittr,"%<>%")
-importFrom(magrittr,"%>%")
-importFrom(magrittr,"%T>%")
-importFrom(rio,import)
-importFrom(stats,na.omit)
-importFrom(stats,t.test)
-importFrom(stringi,stri_detect_fixed)
-importFrom(tidyr,separate_rows)
-importFrom(utils,download.file)
-importFrom(uuid,UUIDgenerate)
-importFrom(vcfR,read.vcfR)
-importFrom(zoo,rollapply)
