+++ R/antigen.garnish_assemble.R
@@ -9,11 +9,18 @@
 
 list_mhc <- function(){
 
-system.file("extdata",
-      "all_alleles.txt", package = "antigen.garnish") %>%
-                      data.table::fread(header = FALSE, sep = "\t") %>%
-                      .$V1
-  }
+  dt <- system.file("extdata",
+                    "all_alleles.txt", package = "antigen.garnish") %>%
+                        data.table::fread(header = FALSE, sep = "\t") %>%
+                        data.table::setnames("V1", "MHC") %>%
+                        .[, species := "human"] %>%
+                        .[MHC %likef% "H-2", species := "mouse"] %>%
+                        .[, class := "II"] %>%
+                        .[MHC %like% "(H-2-[A-Z][a-z])|(HLA-[ABCE]\\*)", class := "I"]
+
+      return(dt)
+
+}
