+++ R/antigen.garnish_pkg.R
@@ -33,6 +33,7 @@
 #' @import parallel
 #' @import stringr
 #' @import testthat
+#' @import knitr
 #' @importFrom rio import
 #' @importFrom ShortRead sread readFasta
 #' @importFrom Biostrings DNAString translate
