+++ inst/extdata/src/main.py
@@ -7,8 +7,9 @@ from Aligner import Aligner
 
 from Neoantigen import Neoantigen
 
-def readNeoantigens(neofilename):
+def readNeoantigens(neofilename, nmerl):
     neoantigens={}
+    nl=nmerl
     f=open(neofilename)
     header=f.readline()
     htab=header.strip().split("\t")
@@ -23,7 +24,7 @@ def readNeoantigens(neofilename):
         if nparams[7]=="NA":
             line=f.readline()
             continue
-        neoantigen=Neoantigen(nparams)
+        neoantigen=Neoantigen(nparams, nl)
         neoantigens[neoantigen.id]=neoantigen
         neoantigen.setA()
         line=f.readline()
@@ -49,7 +50,7 @@ def main(argv):
     outfile=sys.argv[5]
     nmerl=float(argv[6])
 
-    [neoantigens,samples]=readNeoantigens(neofile)
+    [neoantigens,samples]=readNeoantigens(neofile, nmerl)
     #Compute TCR-recognition probabilities for all neoantigens
     aligner=Aligner()
     for sample in samples:
