--- man/garnish_fitness (Andrew Rech's conflicted copy 2018-03-25).Rd
@@ -1,26 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/antigen.garnish_fitness.R
-\name{garnish_fitness}
-\alias{garnish_fitness}
-\title{Generate neoepitope fitness model}
-\usage{
-garnish_fitness(dt, a = 26, k = 4.86936)
-}
-\arguments{
-\item{dt}{Data table output from \code{garnish_predictions}.}
-
-\item{a}{Fitness model parameter. Binding curve horizontal displacement used to determine TCR recognition probability of a peptide compared to an IEDB near match.}
-
-\item{k}{Fitness model parameter. Steepness of the binding curve at \code{a}.}
-}
-\value{
-A data table with added fitness model parameter columns:
-\itemize{
-\item \strong{ResidueChangeClass}: Amino acid change class (e.g. hydrophobic, non-hydrophobic).
-\item \strong{R}: TCR recognition probability, determined by comparison to known epitopes in the IEDB.
-\item \strong{NeoantigenRecognitionPotential}: Neoantigen Recognition Potential calculated by the model. Original validation was limited to 9mers, class I, and human IEDB epitopes. The model will be applied to all peptides, class I and II.
-}
-}
-\description{
-Implements the neoepitope fitness model of \href{https://www.ncbi.nlm.nih.gov/pubmed/29132144}{Luksza et al. Nature 2017}.
-}
