--- man/garnish_predictions (Andrew Rech's conflicted copy 2018-03-25).Rd
@@ -1,192 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/antigen.garnish_predict.R
-\name{garnish_predictions}
-\alias{garnish_predictions}
-\title{Perform neoepitope prediction.}
-\usage{
-garnish_predictions(dt = NULL, path = NULL, assemble = TRUE,
-  generate = TRUE, predict = TRUE, blast = fitness, fitness = TRUE,
-  humandb = "GRCh38", mousedb = "GRCm38", save2wd = FALSE,
-  remove_wt = TRUE)
-}
-\arguments{
-\item{dt}{Data table. Input data table from \code{garnish_variants} or \code{garnish_jaffa}, or a data table in one of these forms:
-
-dt with transcript id:\preformatted{Column name                 Example input
-
-sample_id                   sample_1
-ensembl_transcript_id       ENST00000311936
-cDNA_change                 c.718T>A
-MHC                         HLA-A02:01 HLA-A03:01
-                            H-2-Kb H-2-Kb
-                            HLA-DRB111:07 [second type]
-}
-
-dt with peptide:\preformatted{Column name                 Example input
-
-sample_id                   <same as above>
-pep_mut                     MTEYKLVVVDAGGVGKSALTIQLIQNHFV
-mutant_index                all
-                            7
-                            7 13 14
-MHC                         <same as above>
-}}
-
-\item{path}{Path to input table (\href{https://cran.r-project.org/web/packages/rio/vignettes/rio.html#supported_file_formats}{acceptable formats}).}
-
-\item{assemble}{Logical. Assemble data table?}
-
-\item{generate}{Logical. Generate peptides?}
-
-\item{predict}{Logical. Predict binding affinities?}
-
-\item{blast}{Logical. Run BLASTp to find wild-type peptide and known IEDB matches?}
-
-\item{fitness}{Logical. Run model of \href{https://www.ncbi.nlm.nih.gov/pubmed/29132144}{Luksza et al. Nature 2017} to predict neoepitope fitness?}
-
-\item{humandb}{Character vector. One of "GRCh37" or "GRCh38".}
-
-\item{mousedb}{Character vector. One of "GRCm37" or "GRCm38".}
-
-\item{save2wd}{Logical. Save a copy of garnish_predictions output to the working directory as "ag_output.txt"? Default is \code{FALSE}.}
-
-\item{remove_wt}{Logical. Check all nmers peptides generated against wt peptidome and remove matches. Default is \code{TRUE}. If investigating wild-type sequences, set this to \code{FALSE}.}
-}
-\value{
-A data table of binding predictions including:
-\itemize{
-\item \strong{cDNA_seq}: mutant cDNA sequence
-\item \strong{cDNA_locs}: starting index of mutant cDNA
-\item \strong{cDNA_locl}: ending index of mutant cDNA
-\item \strong{cDNA_type}: netMHC prediction tool output
-\item \strong{frameshift}: frameshift variant?
-\item \strong{coding}: wt cDNA sequence
-\item \strong{coding_mut}: mutant cDNA sequence
-\item \strong{pep_type}: type of peptide
-\item \strong{pep_mut}: mutant peptide sequence
-\item \strong{pep_wt}: wt peptide sequence
-\item \strong{mismatch_s}: starting index of mutant peptide sequence
-\item \strong{mismatch_l}: ending index of mutant peptide sequence
-\item \strong{mutant_index}: index of mutant peptide
-\item \strong{nmer}: nmer for prediction
-\item \strong{nmer_i}: index of nmer in sliding window
-\item \strong{*_net}: netMHC prediction tool output
-\item \strong{mhcflurry_*}: mhcflurry_ prediction tool output
-\item \strong{mhcnuggets_*}: mhcnuggets_ prediction tool output
-\item \strong{Consensus_scores}: Average value of MHC binding affinity from all prediction tools that contributed output. 95\% confidence intervals are given by \strong{Upper_CI}, \strong{Lower_CI}.
-\item \strong{DAI}: Differential agretopicty index of missense and corresponding wild-type peptide, see \code{garnish_summary} for an explanation of DAI.
-\item \strong{BLAST_A}: Ratio of consensus binding affinity of mutant peptide / closest single AA mismatch from blastp results. Returned only if \code{blast = TRUE}.
-\item \strong{iedb_score}: R implementation of TCR recognition probability for peptide through summing of alignments in IEDB for corresponding organism.
-\item \strong{min_DAI}: Minimum of BLAST_A and DAI values, to provide the most conservative \code{fitness_score} value.
-\item \strong{fitness_score}: Product of min_DAI and iedb_score. The peptide with the highest value per sample is the dominant neoepitope.
-}
-
-fitness model information \href{https://www.ncbi.nlm.nih.gov/pubmed/29132144}{Luksza et al. Nature 2017}:
-\itemize{
-\item \strong{ResidueChangeClass}: Amino acid change class (e.g. hydrophobic, non-hydrophobic).
-\item \strong{A}: Component of the fitness model. Differential MHC affinity of mutant and closest wt peptide, equivalent to DAI if available, otherwise uses BLAST_A.
-\item \strong{R}: TCR recognition probability, determined by comparison to known epitopes in the IEDB and amino acid properties.
-\item \strong{NeoantigenRecognitionPotential}: Product of A and R.
-}
-
-transcript description:
-\itemize{
-\item description
-\item start_position
-\item end_position
-\item transcript_end
-\item transcript_length
-\item transcript_start
-\item peptide
-}
-}
-\description{
-Perform ensemble neoepitope prediction on a data table of missense mutations, insertions, deletions or gene fusions using netMHC, mhcflurry, and mhcnuggets.
-See \code{list_mhc} for compatible MHC allele syntax.  Multiple MHC alleles for a single sample_id should be space separated.
-Please keep \strong{murine and human alleles in separate rows}, this will not affect computational speed but will ensure the correct IEDB file is chosen.  See example.
-}
-\examples{
-\dontrun{
-library(magrittr)
-library(antigen.garnish)
-
-# run full pipeline
-
-   # download an example VCF
-   dt <- "antigen.garnish_example.vcf" \%T>\%
-   utils::download.file("http://get.rech.io/antigen.garnish_example.vcf", .) \%>\%
-
-   # extract variants
-   antigen.garnish::garnish_variants \%>\%
-
-   # add space separataed MHC types with antigen.garnish::list_mhc() compatible names
-       .[, MHC := c("HLA-A*02:01 HLA-DRB1*14:67",
-                   "H-2-Kb H-2-IAd",
-                   "HLA-A*01:47 HLA-DRB1*03:08")] \%>\%
-
-   # predict neoepitopes
-   antigen.garnish::garnish_predictions \%>\%
-
-   # summarize predictions
-   antigen.garnish::garnish_summary \%T>\%
-   print
-
-}
-
-\dontrun{
-# input a data table of transcripts
-
-library(magrittr)
-library(antigen.garnish)
-
- dt <- data.table::data.table(
-          sample_id = "test",
-          ensembl_transcript_id =
-          c("ENSMUST00000128119",
-            "ENSMUST00000044250",
-            "ENSMUST00000018743"),
-          cDNA_change = c("c.4988C>T",
-                          "c.1114T>G",
-                          "c.718T>A"),
-          MHC = c("HLA-A*02:01 HLA-DRB1*14:67",
-                  "H-2-Kb H-2-IAd",
-                  "HLA-A*01:47 HLA-DRB1*03:08")) \%>\%
- antigen.garnish::garnish_predictions \%T>\%
- str
-}
-
-\dontrun{
-# input a data table of peptides for all MHC types
-
-library(data.table)
-library(magrittr)
-library(antigen.garnish)
-
- dt <- data.table::data.table(
-          sample_id = "test",
-          pep_mut = "MTEYKLVVVGAGDVGKSALTIQLIQNHFVDEYDP",
-          mutant_index = "12",
-          MHC = "all") \%>\%
- antigen.garnish::garnish_predictions \%T>\%
- str
-}
-
-\dontrun{
-# input from Microsoft Excel
-
-   # download an example Excel file
-     path <- "antigen.garnish_test_input.xlsx" \%T>\%
-       utils::download.file("http://get.rech.io/antigen.garnish_test_input.xlsx", .)
-
-   # predict neoepitopes
-     dt <- antigen.garnish::garnish_predictions(path = path) \%T>\%
-     str
-}
-
-}
-\references{
-Luksza, M, Riaz, N, Makarov, V, Balachandran VP, et al. 2017. A neoepitope fitness model predicts tumour response to checkpoint blockade immunotherapy. Nature. 23;551(7681):512-516
-}
-\seealso{
-\code{\link{garnish_summary}}
-}
