--- man/make_BLAST_uuid (Andrew Rech's conflicted copy 2018-03-25).Rd
@@ -1,14 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/antigen.garnish_predict.R
-\name{make_BLAST_uuid}
-\alias{make_BLAST_uuid}
-\title{Internal function to categorize mutant and wild-type peptides by similarity using \code{BLAST} to calculate neoepitope amplitude and IEDB homology.}
-\usage{
-make_BLAST_uuid(dti)
-}
-\arguments{
-\item{dti}{Data table of nmers.}
-}
-\description{
-Internal function to categorize mutant and wild-type peptides by similarity using \code{BLAST} to calculate neoepitope amplitude and IEDB homology.
-}
